refactor: move create and submit methods to stripe-elements
